main file cleanup, pcapparser cleanup
