+++ .gitignore
@@ -1,3 +1,6 @@
 venv
 __pycache__
 .idea
+config.ini
+csv_files
+*.pcap
