+++ bin/ndpiReader_arch
