+++ config.example.ini
@@ -10,9 +10,9 @@ useTrainedFeatureProcessors = 1
 [parser]
 
 # if packetLimitPerFlow=0, flows are not limited at all 
-PCAPfilename = ../deviceEstimator/traffic_dumps/skypeLANhome.pcap
+PCAPfilename = dump.pcap 
 packetLimitPerFlow = 100
-nDPIfilename = ndpiReader_deb
+nDPIfilename = ./bin/ndpiReader_deb
 
 csvFileTraining = flows_100split_iotMerged.csv
